Fix backwards ZF in 386 simulator
The implementation was setting ZF to the opposite of the correct value,
reversing the sense of every conditional branch in the program.  This is
the bug that was causing tinyboot1 to crash; comparing a simulated
execution trace with a trace run under GDB found this pretty quickly.
Fixing it actually makes it crash sooner, after only some 11700
instructions instead of almost 16000.